pyproject.toml
[project]
name = "langgraph-deepagents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepagents>=0.2.8",
    "fastapi>=0.122.0",
    "langchain-google-genai>=3.2.0",
    "langgraph-cli[inmem]>=0.4.7",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "uvicorn>=0.38.0",
]
